Fix some quality issues
