+++ metrics/mypy_high_water_mark
@@ -1,1 +1,1 @@
-89.6700
+89.7200
