+++ records_mover/db/postgres/copy_options/types.py
@@ -1,0 +1,3 @@
+from typing import Dict
+
+PostgresCopyOptions = Dict[str, object]
