+++ tests/unit/db/postgres/test_postgres_copy_options_load_known.py
@@ -1,6 +1,6 @@
 import unittest
 from records_mover.db.postgres.loader import PostgresLoader
-from records_mover.db.postgres.postgres_copy_options import postgres_copy_options
+from records_mover.db.postgres.copy_options import postgres_copy_options
 from records_mover.records import ProcessingInstructions, DelimitedRecordsFormat
 from records_mover.records.load_plan import RecordsLoadPlan
 from mock import Mock
